+++ src/dolphin/gx/GXTransform.c
@@ -296,7 +296,7 @@ void GXSetScissorBoxOffset(s32 x_off, s32 y_off) {
     y_off += 0x156;
     x_off = ((u32)x_off >> 1) & 0xFFF003FF;
     y_off = ((u32)y_off >> 1) << 10;
-    reg = ((u32)x_off | (u32)y_off) & 0x00FFFFFF;
+    reg = ((u32)x_off | y_off) & 0x00FFFFFF;
     reg |= 0x59000000;
     GX_WRITE_RAS_REG(reg);
     __GXData->bpSentNot = 0;
